+++ src/store/move-to-dialog/move-to-dialog.ts
@@ -1,0 +1,9 @@
+// Copyright (C) The Arvados Authors. All rights reserved.
+//
+// SPDX-License-Identifier: AGPL-3.0
+
+export interface MoveToFormDialogData {
+    name: string;
+    uuid: string;
+    ownerUuid: string;
+}
