+++ src/views-components/context-menu/action-sets/project-action-set.ts
@@ -11,7 +11,7 @@ import { ToggleFavoriteAction } from "../actions/favorite-action";
 import { toggleFavorite } from "~/store/favorites/favorites-actions";
 import { favoritePanelActions } from "~/store/favorite-panel/favorite-panel-action";
 import { PROJECT_CREATE_DIALOG } from "../../dialog-create/dialog-project-create";
-import { openMoveProjectDialog } from '../../move-project-dialog/move-project-dialog';
+import { openMoveProjectDialog } from '~/store/move-project-dialog/move-project-dialog';
 
 export const projectActionSet: ContextMenuActionSet = [[
     {
